完成

+++ src/main/java/com/cs/Public.java
@@ -5,5 +5,5 @@ import java.math.BigDecimal;
 public class Public {
     static final int[] NUMS = {1, 2, 3, 4, 5, 6, 7};
     static final int[] RECORDS = {8, 9, 10, 11, 12, 13, 14};
-    static Integer SUM = 0;
+    static Float SUM = 0.0f;
 }
